refactor(e2e): move each env definition into its own test file
Each environment type now lives alongside the test suite that uses it:
- dockerPermissionEnv + provisioner + Diagnose helpers → docker_permission_test.go
- checkFailureEnv + provisioner + constants → check_failure_test.go
- provisioner.go retains only the shared newBaseEC2Env helper
