+++ test/new-e2e/tests/agent-health/check_failure_test.go
@@ -6,21 +6,46 @@
 package agenthealth
 
 import (
+	_ "embed"
+	"errors"
 	"fmt"
 	"testing"
 	"time"
 
+	"github.com/pulumi/pulumi/sdk/v3/go/pulumi"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
 	"github.com/DataDog/agent-payload/v5/healthplatform"
 
+	"github.com/DataDog/datadog-agent/test/e2e-framework/components/datadog/agentparams"
 	"github.com/DataDog/datadog-agent/test/e2e-framework/testing/components"
 	"github.com/DataDog/datadog-agent/test/e2e-framework/testing/e2e"
+	"github.com/DataDog/datadog-agent/test/e2e-framework/testing/provisioners"
+	"github.com/DataDog/datadog-agent/test/e2e-framework/testing/utils/common"
+	"github.com/DataDog/datadog-agent/test/e2e-framework/testing/utils/e2e/client/agentclient"
 )
 
-// fixedCheckPy is the replacement check that succeeds, used in the Resolution phase.
-// The failing version (brokenCheckPy) is defined in provisioner.go alongside the provisioner.
+//go:embed fixtures/health_platform_agent_config.yaml
+var healthPlatformAgentConfig string
+
+// brokenCheckConf is the check configuration placed in conf.d to activate the check.
+const brokenCheckConf = `
+init_config:
+instances:
+  - {}
+`
+
+// brokenCheckPy is a Python check that always raises — deployed by the provisioner.
+const brokenCheckPy = `
+from datadog_checks.base import AgentCheck
+
+class BrokenCheck(AgentCheck):
+    def check(self, instance):
+        raise Exception("synthetic failure for e2e health platform test")
+`
+
+// fixedCheckPy replaces brokenCheckPy in the Resolution phase.
 const fixedCheckPy = `
 from datadog_checks.base import AgentCheck
 
@@ -28,6 +53,49 @@ class BrokenCheck(AgentCheck):
     def check(self, instance):
         self.gauge("e2e.healthplatform.check_ok", 1)
 `
+
+// ============================================================================
+// Environment definition
+// ============================================================================
+
+type checkFailureEnv struct {
+	RemoteHost *components.RemoteHost
+	Agent      *components.RemoteHostAgent
+	Fakeintake *components.FakeIntake
+}
+
+func checkFailureEnvProvisioner() provisioners.PulumiEnvRunFunc[checkFailureEnv] {
+	return func(ctx *pulumi.Context, env *checkFailureEnv) error {
+		base := &baseEC2Env{
+			RemoteHost: env.RemoteHost,
+			Agent:      env.Agent,
+			Fakeintake: env.Fakeintake,
+		}
+		return newBaseEC2Env(ctx, base, "checkfailurevm",
+			agentparams.WithAgentConfig(healthPlatformAgentConfig),
+			agentparams.WithIntegration("broken_check.d", brokenCheckConf),
+			agentparams.WithFile(
+				"/etc/datadog-agent/checks.d/broken_check.py",
+				brokenCheckPy,
+				true,
+			),
+		)
+	}
+}
+
+var _ common.Diagnosable = (*checkFailureEnv)(nil)
+
+func (e *checkFailureEnv) Diagnose(_ string) (string, error) {
+	if e.Agent == nil {
+		return "", errors.New("agent not initialized")
+	}
+	out := e.Agent.Client.Diagnose(agentclient.WithArgs([]string{"--include", healthIssueSuite}))
+	return "==== agent diagnose health-issues ====\n" + out, nil
+}
+
+// ============================================================================
+// Test suite
+// ============================================================================
 
 type checkFailureSuite struct {
 	e2e.BaseSuite[checkFailureEnv]
@@ -61,23 +129,18 @@ func (suite *checkFailureSuite) TestCheckFailureIssueLifecycle() {
 
 	RunHealthIssueLifecycle(suite.T(),
 		HealthIssueTestCase{
-			// IssueName is matched (substring) against the diagnose entry Name field,
-			// which the checkfailure template populates with the check name.
+			// IssueName is matched (substring) against the diagnose entry Name field.
 			IssueName: "broken_check",
-			IssueID:   "check-execution-failure:broken_check*",
-
-			// TriggerIssue re-deploys the broken check after the Resolution cleanup
-			// restores the working version, so the environment is ready for a re-run.
+			// IssueID ends with "*" for prefix matching: the real ID includes a
+			// runtime-generated hash suffix ("check-execution-failure:broken_check:<hash>").
+			IssueID: "check-execution-failure:broken_check*",
+			// TriggerIssue re-deploys the broken check after the Resolution cleanup.
 			TriggerIssue: func(t *testing.T, h *components.RemoteHost) {
 				writeCheckFile(t, h, brokenCheckPy)
 			},
-
-			// FixIssue replaces the failing check with one that succeeds.
 			FixIssue: func(t *testing.T, h *components.RemoteHost) {
 				writeCheckFile(t, h, fixedCheckPy)
 			},
-
-			// AssertMetadata validates the fakeintake payload metadata fields.
 			AssertMetadata: func(t *testing.T, issue *healthplatform.Issue) {
 				assert.Equal(t, "check_execution_failure", issue.IssueName)
 				assert.Equal(t, "check-execution", issue.Category)
@@ -87,14 +150,13 @@ func (suite *checkFailureSuite) TestCheckFailureIssueLifecycle() {
 				assert.NotEmpty(t, issue.Remediation.Summary)
 			},
 		},
-		agent,
-		host,
-		fi,
+		agent, host, fi,
 	)
 }
 
 // writeCheckFile writes content to /etc/datadog-agent/checks.d/broken_check.py.
-// It writes to a world-writable temp path first, then uses sudo to move the file
+// It writes to a temp path first (SFTP user, no sudo), then sudo-moves it into
+// the protected directory with correct ownership.
 func writeCheckFile(t *testing.T, h *components.RemoteHost, content string) {
 	t.Helper()
 	const (
